third_party/blink/web_tests/wpt_internal/prerender/restriction-download.html

<!DOCTYPE html>
<!--
This file cannot be upstreamed to WPT until:
* startPrerendering() usage is replaced with a WebDriver API
* The test changes to expect the deferring behavior described at
  https://jeremyroman.github.io/alternate-loading-modes/#patch-downloading
  instead of expecting that prerender is cancelled.
-->
<title>
Prerending is canceled upon the `download the hyperlink` algorithm
</title>
<meta name="timeout" content="long">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<body>
<script>

promise_test(async t => {
  const bc = new BroadcastChannel('test-channel');
  t.add_cleanup(_ => bc.close());

  const gotMessage = new Promise(resolve => {
    bc.addEventListener('message', e => {
      resolve(e.data);
    }, {
      once: true
    });
  });
  const url = `resources/download.html`;
  window.open(url, '_blank', 'noopener');

  const result = await gotMessage;
  assert_equals(result, 'prerendering is canceled');
}, `attempting a download should cause prerendering to be cancelled`);

</script>
</body>
